- Fixed the default items to return their natural type in getObject()

+++ plc4j/protocols/driver-bases/base/src/main/java/org/apache/plc4x/java/base/messages/items/DefaultBigDecimalFieldItem.java
@@ -31,7 +31,7 @@ public class DefaultBigDecimalFieldItem extends FieldItem<BigDecimal> {
 
     @Override
     public Object getObject(int index) {
-        return getDouble(index);
+        return getBigDecimal(index);
     }
 
     @Override
