+++ plc4j/protocols/driver-bases/base/src/main/java/org/apache/plc4x/java/base/messages/items/DefaultBigIntegerFieldItem.java
@@ -31,7 +31,7 @@ public class DefaultBigIntegerFieldItem extends FieldItem<BigInteger> {
 
     @Override
     public Object getObject(int index) {
-        return getLong(index);
+        return getBigInteger(index);
     }
 
     @Override
